Add initial draft of `git-branch-list`
A Git add-on that stores a list of branch names and allows you to
quickly switch to any branch in the list.

This enables you to more quickly switch between branches that you're
working on or reviewing without having to either remember or look up the
beginnings of the names of the branches.

Currently provides basic functionality for saving and dropping branches
from the list, printing the list, and checking out a branch in the list.

The branch list "database" is stored in a text file in the `.git/info/`
directory local to the project. Perhaps worrisome to store an external
file in the internal Git directory, but this allows us to keep the
database local to the current project and maintain different databases
for different projects.