git-branch-list(drop_branch): Allow dropping branches by ID
In addition to letting you drop the current branch and a parameter with
the name of the branch, allow branches to be dropped by their ID if they
don't match an existing branch.

Extract the code that finds a branch in the database by its ID to a
function so it can be used by both the `checkout_branch` and
`drop_branch` functions.