git-branch-list: Remove obsolete comments
Some ideas or old versions of code that is no longer relevant. I've also
decided not to implement the TODO and have the list IDs change when you
change the list like in `git stash`, at least for now.

I'll keep the TODO in mind when I use the program and reevaluate to see
whether it makes sense to have fixed IDs.