t/: Move `$BIN` variable to module
In order to be able to easily reuse the `$BIN` variable we created in
the test for the `clear` sub-command, move it to a Perl module that can
be included in other tests.

Add the `t/` directory to the include path when running `prove` to
ensure that `bin.pm` can be found and included.

Thanks to these resources for explaining Perl modules:
https://perlmaven.com/how-to-create-a-perl-module-for-code-reuse
https://stackoverflow.com/questions/23899121/perl-declare-and-export-variables-from-a-module/23900384#23900384
https://stackoverflow.com/questions/17931981/what-is-isa-in-perl/17932340#17932340