git-branch-list(drop): Fix deletion for branches with slashes
Previously, if a branch name had slashes in it, the `drop` command would
fail. This is because it used `sed` with `/` separators, and the
unescaped slashes in the branch name would mess up sed's pattern.

Change this to use grep instead of sed so we don't have to worry about
escaping and special characters in branch names.