git-branch-list: Fix checkout and drop commands for >10 saved branches
Previously there was a bug when you had 11 or more saved branches. Saved
branches 1 and 11 would both get matched by the regex in `branch_by_id`,
causing the two branches to be concatenated.

This resulted in a checkout error, preventing checkout. When dropping a
branch, both in the pair would be dropped, even if only branch #1 was
specified.

Surround the branch IDs with whitespace to prevent incorrect matches.